Fix copy handler preference text to config
The user is no longer responsible for keeping hold of the config preference string.
That was an oversight on my part and has now been fixed.
To accommodate this, the vb_config_initialise function now takes two parameters.
The preference string given to the second parameter is copied into the config->handler_preference string.